Generated by Spark: https://www.uxtools.co/ https://www.uxtools.co/ http://designer.tips/ https://www.youtube.com/watch?v=1i8s8knHFTs&t=155s https://www.youtube.com/watch?v=Qj1FK8n7WgY&t=37s
Utilize the above to understand how to create the best possible layout.

Also I definitely asked you to work on more games then that please ensure that all the games and mini-games are all working with the proper game mechanics. All that you leasrned in the links I gave you please use the principles to craft the best possible games. Thanks!